pyproject.toml
[project]
name = "jirato"
version = "0.1.0"
description = "JIRA automation tool"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "loguru>=0.7.0",
    "jupyterlab>=4.4.4",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "ollama>=0.2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["jirato"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]
